HITL: cycle power for "QUSB" port (#75)
uhubctl supports a "cycle" action and can filter by the device name. If
the device isn't found we try to recover by turning on power to all
ports on the default hub for the Pi.

I tried filtering by `-n 05c6` which I thought would be more reliable,
but it didn't seem to work, maybe that's filtering the USB hub vendor
instead